+++ app/decorators/resident_decorator.rb
@@ -1,6 +1,10 @@
 class ResidentDecorator < Draper::Decorator
   delegate_all
 
+  def icon_name
+    'user'
+  end
+
   def property_summary(type: :icons)
     h.render 'properties/property_icon_list', properties: object.properties.decorate
   end